fix(ui): move 17 misplaced test files from src/primitives to test/primitives
Tests belong in the mirror test/ tree, not alongside source. These files
in src/primitives/ caused the registry to serve .test.ts files as
primitives to consumers.

15 unique files moved with import paths rewritten from ./ to
../../src/primitives/. 2 duplicates (collision-detector, history) had
their unique coverage (point-anchor kinds, reactive snapshot API) merged
into the canonical test/ versions. Fixture paths updated for
token-sheet and serializer-mdx-fixtures.
