+++ packages/ui/test/primitives/serializer-mdx-fixtures.test.ts
@@ -1,16 +1,16 @@
 import { readFileSync } from 'node:fs';
 import { join } from 'node:path';
 import { describe, expect, it } from 'vitest';
-import type { SerializerBlock } from './serializer';
-import { mdxSerializer } from './serializer-mdx';
-import type { InlineContent } from './types';
+import type { SerializerBlock } from '../../src/primitives/serializer';
+import { mdxSerializer } from '../../src/primitives/serializer-mdx';
+import type { InlineContent } from '../../src/primitives/types';
 
 // =============================================================================
 // Helpers
 // =============================================================================
 
 function loadFixture(name: string): string {
-  return readFileSync(join(__dirname, '__fixtures__', name), 'utf-8');
+  return readFileSync(join(__dirname, '../../src/primitives/__fixtures__', name), 'utf-8');
 }
 
 function deserialize(input: string) {
